services/payment-service/requirements.txt
aiohttp
azure-functions
